requirements.txt
APScheduler>=3.10.0
mysql-connector-python>=8.3.0
